+++ automation/google/google_rbd_from_label/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.2
+
+- fixed link to README in policy description
+
 ## v1.1
 
 - added ability to specify names for the newly created dimensions