+++ cost/aws/savings_plan/expiration/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v2.1
+
+- fixed link to README in policy description
+
 ## v2.0
 
 - Initial Release