+++ cost/turbonomics/credential_refresh/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.4
+
+- fixed link to README in policy description
+
 ## v0.3
 
 - Bug fix: Previously this policy could only update applied policies (policies added using the Automation > Templates panel) but no aggregated policies like those executed from the Automation > Catalog, now this was fixed and the policy can work with both.